rename to lidNick

+++ src/Glint.Networking/Game/Updates/GameUpdate.cs
@@ -3,7 +3,7 @@ using Glint.Networking.Messages;
 namespace Glint.Networking.Game.Updates {
     public abstract class GameUpdate {
         public long time;
-        public uint sourceUid;
+        public long sourceUid;
 
         public virtual void copyFrom(GameUpdateMessage msg) {
             time = msg.time;
