+++ src/Glint.Networking/Messages/GameUpdateMessage.cs
@@ -6,7 +6,7 @@ using MsgPack.Serialization;
 namespace Glint.Networking.Messages {
     public abstract class GameUpdateMessage : LimeMessage {
         [MessagePackMember(0)] public long time { get; set; }
-        [MessagePackMember(1)] public uint sourceUid { get; set; }
+        [MessagePackMember(1)] public long sourceUid { get; set; }
 
         public override NetDeliveryMethod deliveryMethod => NetDeliveryMethod.UnreliableSequenced;
